Moved NodeDataGen to pl.touk.nussknacker.test.utils.generators in designer
